Moving hooks to Classes/Hooks and create utilities for Hooks and Scheduler tasks
