+++ .scrutinizer.yml
@@ -2,11 +2,10 @@ filter:
     excluded_paths:
         - 'tests/*'
     paths:
+        - 'Classes/*'
         - 'cli/*'
         - 'domain/*'
-        - 'hooks/*'
         - 'modfunc1/*'
-        - 'scheduler/*'
         - 'system/*'
         - 'template/*'
         - 'view/*'
