+++ .travis.yml
@@ -72,13 +72,13 @@ script:
     if [[ "$COVERAGE" == "0" ]]; then
       echo;
       echo "Running unit tests";
-      find 'Tests/Unit' -name '*testcase.php' | xargs .Build/bin/phpunit --colors -c $UNIT_XML
+      .Build/bin/phpunit --colors -c $UNIT_XML Tests/Unit
     fi
   - >
     if [[ "$COVERAGE" == "1" ]]; then
       echo;
       echo "Running unit tests";
-      find 'Tests/Unit' -name '*testcase.php' | xargs .Build/bin/phpunit --coverage-clover=unittest-coverage.clover --colors -c $UNIT_XML
+      .Build/bin/phpunit --coverage-clover=unittest-coverage.clover --colors -c $UNIT_XML Tests/Unit
     fi
   - >
     if [[ "$COVERAGE" == "1" ]]; then
@@ -95,7 +95,7 @@ script:
       export typo3DatabaseHost="localhost";
       export typo3DatabaseUsername="root";
       export typo3DatabasePassword="";
-      find 'Tests/Functional' -wholename '*testcase.php' | parallel --gnu 'echo; echo "Running functional test suite {}"; .Build/bin/phpunit --coverage-clover={}functionaltest-coverage.clover --colors  -c $FUNCTIONAL_XML {}'
+      find 'Tests/Functional' -wholename '*testcase.php' -o -name '*Test.php' | parallel --gnu 'echo; echo "Running functional test suite {}"; .Build/bin/phpunit --coverage-clover={}functionaltest-coverage.clover --colors  -c $FUNCTIONAL_XML {}'
     fi
   - >
     echo;
