+++ Tests/Unit/Hooks/ProcessCleanUpHookTest.php
@@ -1,0 +1,78 @@
+<?php
+namespace AOE\Crawler\Tests\Unit\Hooks;
+
+/***************************************************************
+ *  Copyright notice
+ *
+ *  (c) 2016 AOE GmbH <dev@aoe.com>
+ *
+ *  All rights reserved
+ *
+ *  This script is part of the TYPO3 project. The TYPO3 project is
+ *  free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  The GNU General Public License can be found at
+ *  http://www.gnu.org/copyleft/gpl.html.
+ *
+ *  This script is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  This copyright notice MUST APPEAR in all copies of the script!
+ ***************************************************************/
+
+use TYPO3\CMS\Core\Tests\UnitTestCase;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+use AOE\Crawler\Hooks\ProcessCleanUpHook;
+
+/**
+ * Class ProcessCleanUpHookTest
+ * @package AOE\Crawler\Tests\Unit\Hooks
+ */
+class ProcessCleanUpHookTest extends UnitTestCase
+{
+
+    /**
+     * @var \AOE\Crawler\Hooks\ProcessCleanUpHook
+     * @inject
+     */
+    var $subject = null;
+
+    public function setUp()
+    {
+        parent::setUp();
+        /** @var ProcessCleanUpHook subject */
+        //$this->subject = GeneralUtility::makeInstance('AOE\\Crawler\\Hooks\\ProcessCleanUpHook');
+    }
+
+    public function tearDown()
+    {
+        unset($this->subject);
+        parent::tearDown();
+    }
+
+    /**
+     * @test
+     */
+    public function tempTest()
+    {
+        $this->assertTrue(true);
+    }
+
+    /**
+     * test
+     */
+    public function createResponseArrayReturnsEmptyArray()
+    {
+        $emptyInputString = '';
+
+        $this->assertEquals(
+            array(),
+            $this->callInaccessibleMethod($this->subject,'createResponseArray', $emptyInputString)
+        );
+    }
+}
