+++ ext_autoload.php
@@ -1,7 +1,4 @@
 <?php
-// DO NOT CHANGE THIS FILE! It is automatically generated by extdeveval::buildAutoloadRegistry.
-// This file was generated on 2009-12-21 14:12
-
 $extensionPath = \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extPath('crawler');
 return array(
 	'tx_crawler_api' => $extensionPath . 'class.tx_crawler_api.php',
@@ -26,9 +23,6 @@ return array(
 	'tx_crawler_hooks_staticFileCacheCreateUri' => $extensionPath . 'hooks/class.tx_crawler_hooks_staticFileCacheCreateUri.php',
 	'tx_crawler_hooks_processCleanUp' => $extensionPath . 'hooks/class.tx_crawler_hooks_processCleanUp.php',
 	'tx_crawler_modfunc1' => $extensionPath . 'modfunc1/class.tx_crawler_modfunc1.php',
-	'tx_crawler_api_testcase' => $extensionPath . 'tests/class.tx_crawler_api_testcase.php',
-	'tx_crawler_domain_events_dispatcher_testcase' => $extensionPath . 'tests/events/class.tx_crawler_domain_events_dispatcher_testcase.php',
-	'tx_crawler_domain_events_test' => $extensionPath . 'tests/events/data/class.tx_crawler_domain_events_test.php',
 	'tx_crawler_view_pagination' => $extensionPath . 'view/class.tx_crawler_view_pagination.php',
 	'tx_crawler_view_process_list' => $extensionPath . 'view/process/class.tx_crawler_view_process_list.php',
 );
